Changed '1 hr' to 'hr' + bug fixes.

+++ lib/features/sellAndBuy/screens/checkout_screen.dart
@@ -11,7 +11,6 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:active_ecommerce_flutter/my_theme.dart';
-import 'package:active_ecommerce_flutter/drawer/drawer.dart';
 import 'package:active_ecommerce_flutter/helpers/shared_value_helper.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
@@ -72,7 +71,7 @@ class _CheckoutScreenState extends State<CheckoutScreen> {
     OrderDocument orderDocument = OrderDocument(
       buyerID: cartDoc.data()!['buyer'],
       timestamp: cartDoc.data()!['orderDate'],
-      totalAmount: cartDoc.data()!['totalAmount'],
+      totalAmount: cartDoc.data()!['totalAmount'].toDouble(),
       status: cartDoc.data()!['status'],
       sellers: cartDoc.data()!['sellers'],
       orderItems: orderItems,
